refactor(core-nodes): rename fake_data_generator module to traffic_generator
Aligns the receiver's module folder and Rust symbols with its URN id
(urn:otel:receiver:traffic_generator), matching the convention used by
every other receiver/processor/exporter in core-nodes where the module
name matches the URN id segment.

- crates/core-nodes/src/receivers/fake_data_generator/ -> traffic_generator/
- OTAP_FAKE_DATA_GENERATOR_URN -> TRAFFIC_GENERATOR_RECEIVER_URN
- OTAP_FAKE_DATA_GENERATOR (factory) -> TRAFFIC_GENERATOR_RECEIVER
- FakeGeneratorReceiver -> TrafficGeneratorReceiver
- Updates use paths in validation, otap tests, benchmarks, and READMEs

Intentionally preserved (observability surface stability):
- URN string "urn:otel:receiver:traffic_generator" (already correct)
- Metric set name 'fake_data_generator.receiver.metrics'
- otel_info! log event names 'fake_data_generator.{shutdown,drain_ingress}'
- FakeSignalReceiverMetrics struct (coupled to its metric_set name)
- Semconv scope strings and YAML config file names
- Test helper names (fake_receiver_*)
