+++ rust/otap-dataflow/crates/core-nodes/src/receivers/traffic_generator/README.md
@@ -26,7 +26,7 @@ don't need multiple load-generator cores.
 ```bash
 cd rust/otap-dataflow
 cargo build --release
-bash crates/core-nodes/src/receivers/fake_data_generator/bench/bench.sh
+bash crates/core-nodes/src/receivers/traffic_generator/bench/bench.sh
 ```
 
 ---
