+++ rust/otap-dataflow/crates/core-nodes/src/receivers/traffic_generator/attributes.rs
