+++ rust/otap-dataflow/crates/core-nodes/src/receivers/traffic_generator/bench/sender-static-fresh.yaml
