+++ rust/otap-dataflow/crates/core-nodes/src/receivers/traffic_generator/bench/sut-otlp-forward.yaml
