+++ rust/otap-dataflow/crates/core-nodes/src/receivers/traffic_generator/semconv_signal.rs
@@ -7,8 +7,8 @@
 //!
 //! ToDo: Add profile signal support -> update the builder lib.rs to work on profile object
 
-use crate::receivers::fake_data_generator::attributes::get_attribute_name_value;
-use crate::receivers::fake_data_generator::fake_data::{
+use crate::receivers::traffic_generator::attributes::get_attribute_name_value;
+use crate::receivers::traffic_generator::fake_data::{
     current_time, delay, gen_span_id, gen_trace_id, get_scope_name, get_scope_version,
 };
 use otap_df_pdata::proto::opentelemetry::{
