+++ rust/otap-dataflow/crates/otap/tests/core_node_liveness_tests.rs
@@ -18,7 +18,7 @@ use otap_df_config::policy::{ChannelCapacityPolicy, TelemetryPolicy};
 use otap_df_config::{DeployedPipelineKey, PipelineGroupId, PipelineId};
 use otap_df_core_nodes::processors::batch_processor::OTAP_BATCH_PROCESSOR_URN;
 use otap_df_core_nodes::processors::retry_processor::RETRY_PROCESSOR_URN;
-use otap_df_core_nodes::receivers::fake_data_generator::OTAP_FAKE_DATA_GENERATOR_URN;
+use otap_df_core_nodes::receivers::traffic_generator::TRAFFIC_GENERATOR_RECEIVER_URN;
 use otap_df_engine::context::ControllerContext;
 use otap_df_engine::control::{
     RuntimeControlMsg, pipeline_completion_msg_channel, runtime_ctrl_msg_channel,
@@ -83,7 +83,7 @@ fn build_retry_pipeline_config(
     PipelineConfigBuilder::new()
         .add_receiver(
             "fake_receiver",
-            OTAP_FAKE_DATA_GENERATOR_URN,
+            TRAFFIC_GENERATOR_RECEIVER_URN,
             Some(fake_receiver_config(6, 2, true)),
         )
         .add_processor(
@@ -119,7 +119,7 @@ fn build_batch_pipeline_config(
     PipelineConfigBuilder::new()
         .add_receiver(
             "fake_receiver",
-            OTAP_FAKE_DATA_GENERATOR_URN,
+            TRAFFIC_GENERATOR_RECEIVER_URN,
             Some(fake_receiver_config(3, 3, true)),
         )
         .add_processor(
@@ -157,7 +157,7 @@ fn build_otlp_batch_local_wakeup_pipeline_config(
     PipelineConfigBuilder::new()
         .add_receiver(
             "fake_receiver",
-            OTAP_FAKE_DATA_GENERATOR_URN,
+            TRAFFIC_GENERATOR_RECEIVER_URN,
             Some(rate_limited_fake_receiver_config(5, 1, 1, true)),
         )
         .add_processor(
